feat: Add `markdown-it` plugin to render Markdown to HTML
This commit creates a Nunjucks filter to render markdown. This
will be useful for using the Markdown content in the HTML template.
As part of this commit, I'm adding a TypeScript definition file
for the plugins so the TS compiler will stop complaining. The
plugins need to be in JavaScript so they can be run directly by
Node for pre-compiling the Nunjucks templates.